+++ src/pages/dashboard/index.tsx
@@ -1,8 +1,8 @@
 import { Box, Flex, SimpleGrid, Text, theme } from "@chakra-ui/react";
 import { ApexOptions } from "apexcharts";
 import dynamic from 'next/dynamic';
-import { Header } from "../components/Header";
-import { SideBar } from "../components/Sidebar";
+import { Header } from "../../components/Header";
+import { SideBar } from "../../components/Sidebar";
 
 const Chart = dynamic(() => import('react-apexcharts'), {
   ssr: false
